requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.3.1
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-lru=2.0.5=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bert-score=0.3.13=pypi_0
bertopic=0.17.0=pypi_0
bleach=6.2.0=pypi_0
brotli=1.1.0=hd74edd7_2
brotli-bin=1.1.0=hd74edd7_2
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=py313h0ebd0e5_0
cycler=0.12.1=pyhd8ed1ab_1
debugpy=1.8.14=py313h928ef07_0
decorator=5.2.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
fastjsonschema=2.21.1=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.57.0=py313ha9b7d5b_0
fqdn=1.5.1=pypi_0
freetype=2.13.3=h1d14073_0
fsspec=2025.3.2=pypi_0
h11=0.14.0=pypi_0
hdbscan=0.8.40=pypi_0
httpcore=1.0.8=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
ipykernel=6.29.5=pyh57ce528_0
ipython=9.1.0=pyhfb0248b_0
ipython_pygments_lexers=1.1.1=pyhd8ed1ab_0
ipywidgets=8.1.6=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
joblib=1.4.2=pypi_0
json5=0.12.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-events=0.12.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.15.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
jupyterlab=4.4.0=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.14=pypi_0
kiwisolver=1.4.7=py313hf9c7212_0
krb5=1.21.3=h237132a_0
lcms2=2.17=h7eeda09_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=31_h10e41b3_openblas
libbrotlicommon=1.1.0=hd74edd7_2
libbrotlidec=1.1.0=hd74edd7_2
libbrotlienc=1.1.0=hd74edd7_2
libcblas=3.9.0=31_hb3479ef_openblas
libcxx=20.1.2=ha82da77_0
libdeflate=1.23=hec38601_0
libedit=3.1.20250104=pl5321hafb1f1b_0
libexpat=2.7.0=h286801f_0
libffi=3.4.6=h1da3d7d_1
libgfortran=14.2.0=heb5dd2a_105
libgfortran5=14.2.0=h2c44a93_105
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=31_hc9a63f6_openblas
liblzma=5.8.1=h39f12f2_0
libmpdec=4.0.0=h99b78c6_0
libopenblas=0.3.29=openmp_hf332438_0
libpng=1.6.47=h3783ad8_0
libsodium=1.0.20=h99b78c6_0
libsqlite=3.49.1=h3f77e49_2
libtiff=4.7.0=h551f018_3
libwebp-base=1.5.0=h2471fea_0
libxcb=1.17.0=hdb1d25a_0
libzlib=1.3.1=h8359307_2
llvm-openmp=20.1.2=hdb05f8b_0
llvmlite=0.44.0=pypi_0
loadenv=0.1.1=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=py313h39782a4_0
matplotlib-base=3.10.1=py313haaf02c0_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mistune=3.1.3=pypi_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=pyh9f0ad1d_0
narwhals=1.35.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h5e97a16_3
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
notebook=7.4.0=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.61.2=pypi_0
numpy=2.2.4=py313h41a2e72_0
openai=1.73.0=pypi_0
openjpeg=2.5.3=h8a3d83b_0
openssl=3.5.0=h81ee809_0
overrides=7.7.0=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py313h668b085_3
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=py313hb37fac4_0
pip=25.0.1=pyh145f28c_0
platformdirs=4.3.7=pyh29332c3_0
plotly=6.0.1=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.50=pyha770c72_0
psutil=7.0.0=py313h90d716c_0
pthread-stubs=0.4=hd74edd7_1002
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.22=pypi_0
pydantic=2.11.3=pypi_0
pydantic-core=2.33.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pynndescent=0.5.13=pypi_0
pyparsing=3.2.3=pyhd8ed1ab_1
python=3.13.3=h81fe080_100_cp313
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.1.0=pyh29332c3_1
python-json-logger=3.3.0=pypi_0
python-tzdata=2025.2=pyhd8ed1ab_0
python_abi=3.13=6_cp313
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=pypi_0
pyzmq=26.4.0=py313he6960b1_0
qhull=2020.2=h420ef59_5
readline=8.2=h1d1bf99_2
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.24.0=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=py313h9a24e0a_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
send2trash=1.8.3=pypi_0
sentence-transformers=4.0.2=pypi_0
setuptools=78.1.0=pypi_0
six=1.17.0=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=py313h93df234_0
sympy=1.13.1=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.6.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.13=h5083fa2_1
tokenizers=0.21.1=pypi_0
torch=2.6.0=pypi_0
tornado=6.4.2=py313h90d716c_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.51.2=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-inspection=0.4.0=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
umap-learn=0.5.7=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
widgetsnbextension=4.0.14=pypi_0
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
zeromq=4.3.5=hc1bb282_7
zipp=3.21.0=pyhd8ed1ab_1
zstd=1.5.7=h6491c7d_2
